reused subsample for gdp1h and gdp6h
